+++ src/shared_resources.hpp
@@ -20,10 +20,12 @@
 #include <chrono>
 #include <algorithm>
 #include <atomic>
+#include <deque>
 
 #include "VideoMasterHD_Core.h"
 
 #include "signal_information.hpp"
+#include "rx_renderer.hpp"
 
 namespace Deltacast
 {
@@ -35,27 +37,24 @@ namespace Deltacast
             std::atomic_bool stop_is_requested = false;
             std::atomic_bool signal_has_changed = false;
 
-            bool wait_until_ready_to_process();
-            void notify_processing_finished();
-            bool wait_until_processed();
-            void notify_ready_to_process();
+            void reset();
 
-            std::unique_lock<std::mutex> lock();
+            HANDLE pop_buffer_for_processing();
+            void push_buffer_for_processing(HANDLE buffer);
+
+            HANDLE get_buffer_to_transfer();
+            void set_buffer_to_transfer(HANDLE buffer);
 
         private:
             std::mutex mutex;
             std::condition_variable condition_variable;
 
-            bool to_process = false;
-            bool processed = false;
+            std::deque<HANDLE> buffers_ready_for_processing;
+            std::deque<HANDLE> buffers_ready_for_transfer;
         } synchronization;
 
-        UBYTE* buffer = nullptr;
-        ULONG buffer_size = 0;
-
         SignalInformation signal_info;
-
-        unsigned int maximum_latency;
+        std::optional<RxRenderer> rx_renderer;
 
         void reset();
     };